Switching the order of the Download/Uploaded for remote devices.
I saw people asking for the same thing in issues #821 and #844 and #921 and thought I'd just create a PR that we can debate over instead of merely having a debate.

This feels very unintuitive right now.  I think people will continue to be confused unless there is a change in terminology, iconography, or perhaps the simple code change that I am proposing.